Final fix for the no multiple alerts system
There was another (and hopefully, the last one) flaw in the system we have built up to prevent multiple alerts to be generated.

If moderating an user's thread multiple times, it wouldn't deliver any alert except for the first one. Flushing out the "cache" in the user's session before outputting something to the user is the permanent fix to get it working like a charm.