Fix #293: The "Who's online" location is broken (#294)
Resolves #293

Supersedes #266 (There's nothing especially wrong with that PR except
for its minor failure to remove a mistaken define of NO_ONLINE; I just
happened to implement my own fix in the course of all the other fixes
I'm submitting tonight).

This PR honours the comment @euantorano made
[here](https://github.com/MyBBStuff/MyAlerts/pull/266#issuecomment-593251171)
indicating that a location should be set for the modal. There is, then,
never any reason to define NO_ONLINE for any functionality of this
plugin.